+++ collective/contentrules/mailtogroup/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: UTF-8 -*-
-from zope.i18nmessageid import MessageFactory
-
-
-mailtogroupMessageFactory = MessageFactory('collective.contentrules.mailtogroup')