+++ src/collective/contentrules/mailtogroup/__init__.py
@@ -1,0 +1,4 @@
+from zope.i18nmessageid import MessageFactory
+
+
+mailtogroupMessageFactory = MessageFactory("collective.contentrules.mailtogroup")