+++ src/collective/contentrules/mailtogroup/setuphandlers.py
@@ -1,0 +1,19 @@
+from plone.base.interfaces.installable import INonInstallable
+from zope.interface import implementer
+
+
+@implementer(INonInstallable)
+class HiddenProfiles:
+    def getNonInstallableProfiles(self):
+        """Hide uninstall profile from site-creation and quickinstaller."""
+        return ["collective.contentrules.mailtogroup:uninstall"]
+
+
+def post_install(context):
+    """Post install script"""
+    # Do something at the installation of this package.
+
+
+def uninstall(context):
+    """Uninstall script"""
+    # Do something at the end of the uninstallation of this package.